+++ src/bin/demo.rs
@@ -2,7 +2,8 @@ extern crate silk;
 
 use silk::historian::Historian;
 use silk::log::{verify_slice, Entry, Sha256Hash};
-use silk::event::{generate_keypair, get_pubkey, sign_claim_data, Event};
+use silk::transaction::{generate_keypair, get_pubkey, sign_claim_data};
+use silk::event::Event;
 use std::thread::sleep;
 use std::time::Duration;
 use std::sync::mpsc::SendError;
