+++ src/historian.rs
@@ -7,7 +7,8 @@ use std::sync::mpsc::{sync_channel, Receiver, SyncSender};
 use std::time::Instant;
 use log::{hash, Entry, Sha256Hash};
 use logger::{ExitReason, Logger};
-use event::{Event, Signature};
+use transaction::Signature;
+use event::Event;
 use serde::Serialize;
 use std::fmt::Debug;
 
@@ -68,6 +69,7 @@ mod tests {
     use super::*;
     use log::*;
     use event::*;
+    use transaction::*;
     use std::thread::sleep;
     use std::time::Duration;
