+++ src/lib.rs
@@ -2,6 +2,7 @@
 pub mod log;
 pub mod logger;
 pub mod event;
+pub mod transaction;
 pub mod genesis;
 pub mod historian;
 pub mod accountant;
